perfect script.
