+++ output.csv
@@ -1,12 +1,12 @@
 Serial Number,Company Name:,Member No:,Category:,Year of Established:,Address:,Phone:,Fax:,Email:,Website:,Executive:,Mobile:,Product:,Rawmaterial:
-1300,BAPA SITARAM ICE,B-014/I/L,Industrial Ice,N/A,GIDCPlot No. 25-26/20Phase 1VatvaAhmedabad - 382445,25831889,N/A,steelkingaccessories66@gmail.com / contact@riddhicorporation.com,www.riddhicorporation.com,Harish M. Jajal,9426376861,Industrial Ice,"Ammonia, Salt,, Salt,  Water"
-1301,BARDIA PRODUCTS,B-071/I/L,Food and Food Products,N/A,GIDCPlot No. 77/7Phase 1VatvaAhmedabad - 382445,22910532,N/A,bardiaproducts@yahoo.com,www.bardiaproducts.com,Vimal Bardia,9327021393,"Food (Food Processing), Wheat Rice, Corn",N/A
-1302,BASSOON RUBBER PVT. LTD.,B-016/IV/L,Rubber and Rubber Products,N/A,GIDCPlot No. 4613Phase 4VatvaAhmedabad - 382445,N/A,N/A,mukesh_contact@yahoo.com,N/A,Mukesh Shah,9327004435,Rubber Moulding Articles,Rubber Chemicals
-1303,BENY DYE CHEM,B-017/I/L,Dyes and Dyes Intermediates,N/A,GIDCPlot No. 80/7Phase 1VatvaAhmedabad - 382445,N/A,N/A,N/A,N/A,V. B. Shah,9427116153,Dyes,N/A
-1304,BENZO PRODUCTS (INDIA) PVT. LTD.,B-018/IV/L,Dyes and Dyes Intermediates,1997,GIDCPlot No. 3204Phase 4VatvaAhmedabad - 382445,40084206,N/A,cmd@benzoproducts.com,www.benzoproducts.in,Ramesh P. Patel,9898567676,Rare Dyes Intermediates,N/A
-1305,BHAGAVATI RUBBER PRODUCT INDUSTRIES,B-019/II/A,Rubber and Rubber Products,1982,GIDCPlot No. 411/2Phase 2VatvaAhmedabad - 382445,"25831237, 25895991",26853390,kamlesh.joshi@bhagvati.in,N/A,Dinesh I. Gawarwala,9879106662,Automotive Rubber Flap,Natural Rubber
-1306,BHAGVATI CHEMICAL INDUSTRIES,B-020/II/L,Dyes and Dyes Intermediates,1986,GIDCPlot No. 187Phase 2VatvaAhmedabad - 382445,40190993,N/A,stardyes@stardyes.com,N/A,Govind bhai,9825038280,"Direct Dyes, Acid Dyes",N/A
-1307,BHAGWATI CHEMICALS,B-021/II/L,Dyes and Dyes Intermediates,1955,GIDCPlot No. 5631Phase 2VatvaAhmedabad - 382445,40033940,25834567,bhagwatichemicals5631@gmail.com,N/A,Shailesh K. Patel,9824051541,"Beta Chloro Ethylamine,HCL","Mono ethanolamine, Sulphur Powder, Chlorine Gas"
-1308,BHAGWATI DYES & CHEMICAL INDUSTRIES,B-052/II/A,Pigment,N/A,GIDCPlot No. 40Phase 2VatvaAhmedabad - 382445,65138150,N/A,palash_1024@yahoo.co.in,www.bhagwatipigments.com,Palash Bajaj,9909249884,"Naphthols Azo Pigments Powers : Red, Yellow,  Orange Pigment Emulsions","Bon Acid, Wilineoic, PACA, MNA, DCB, PMP"
-1309,BHAGWATI DYESTUFF INDUSTRIES,B-063/II/L,Dyes and Dyes Intermediates,2005,GIDCPlot No. 433Phase 2VatvaAhmedabad - 382445,40083067,N/A,bhagwatidyestuff@yahoo.in,N/A,Shivemurti B. Yadav,9898044410,"Reactive Dyes, Ramazole Dyes","H Acid, Vs, J Acid"
-1310,BHAGWATI INDUSTRIES,B-024/II/L,Pharmaceuticals and Bulk Drugs,1999,GIDCPlot No. 5630Phase 2VatvaAhmedabad - 382445,40033940,25834567,bhagwati5630@yahoo.com,N/A,Shailesh K. Patel,9824051541,Hydro Benzene,"Caustic Flakes, Formaldehyde, Methanol, Nitrobenzene, Dichlone"
+1400,CREATIVE ENGINEERS,C-073/IV/ASSO/L,Pump and Pump Parts,N/A,GIDC.Plot No.C-1/4317Phase 4VatvaAhmedabad - 382445,"25431453, 25841112",25431454,info@sspumpsindia.com / malharpump@gmail.com,www.sspumpsindia.com,Gaurang B. Shah,9825065939 / 8140179903,"Centrifugal Pump, Self Printing Pump,Barrel Pump, Chemical Process Pump, Mud Pump, Slurry Pump, Air Operate Double Diaphargram Pump, Multi Stage Pumps",N/A
+1401,CRYSTAL ALLOYS & MFG. CO.,C-069/l/L,Steel Mfg and Suppliers,N/A,GIDC.Plot No. 70/1-BPhase 1VatvaAhmedabad - 382445,"25893351, 25893352",N/A,crystalfelry@gmail.com,www.crystalalloys.com,Sanjeev P. Saggi,9825033996,"Ferro Alloys, Foundry Materials",N/A
+1402,CRYSTAL QUINONE PVT. LTD.,C-070/II/L,Dyes and Dyes Intermediates,1983,GIDC.Plot No. C1B-159 & 163Phase 2VatvaAhmedabad - 382445,49069000,40083390,info@crystalquinone.com,www.crystalquinone.com,Manish Sevak | Aniket Shah,9099918990 | 9913608080,"BHA, TBHQ (Antioxident) Hydroquinone","Hydroquinone, Toluene"
+1403,CRYSTAL QUINONE PVT. LTD.,C-055/ll/L,Dyes and Dyes Intermediates,1983,GIDC.Plot No.143 & 145Phase 2VatvaAhmedabad - 382445,49069000,40083390,info@crystalquinone.com,www.crystalquinone.com,Girish Patel | Anang Shah,9824023822 | 9825006600,"2:5 Di Methoxy Amiline, Para Cresidine, DMAVS, PCVS, Anthranilic Acid","Hydroquinone, Para, Cresole, Hydrogen Chloro Sulphonic Acid, Caustic Lye, DMS, Nitric Acid"
+1404,CRYSTAL QUINONE PVT. LTD.,C-056/II/L,Dyes and Dyes Intermediates,1997,"GIDC.Plot No.A/2/20, 21Phase 2VatvaAhmedabad - 382445",49069000,40083390,info@crystalquinone.com,www.crystalquinone.com,Ramsing Chavada | Aniket Shah,9099918996 | 9913608080,"Vinyl  Sulphone, Ester 2:5 DMA, Vinyl Sulphone, Ester Para Cresidine","2:5 DMA, Para, Cresidine, Caustic Lye, Chloro Sulphonic Acid"
+1405,D J K INDUSTRIES,D-001/IV/A,Dyes and Dyes Intermediates,1992,GIDC.2912Phase 4VatvaAhmedabad - 382445,N/A,N/A,dineshpatel@djkindia.com,www.djkindia.com,Dinesh Patel,9825024206,"Heating, Ventilation & Air Conditining Equipment","GP Sheet, Aluminium Section, SS 304, Copper Tube, Aluminium Coil, MS & CRCA Sheets"
+1406,DARLING MUESCO (INDIA) PVT.LTD.,D-002/I/A,Plastic,1964,GIDC.Plot No. 97/APhase 1VatvaAhmedabad - 382445,"25893791, 25893792",25834392,sales@darlingmuesco.com,www.darlingmuesco.com,Kiran H. Shah | K. K. Varghese,9825061383 | 9879430284,"Automatic Control Valve, Safety Valve, Pressure Reduction Station Industrial Valve","Casting, SS Material, MS Material, Spring, Rubber Parts"
+1407,DARSHINI ENGINEERS,D-004/IV/A,Engineering,1999,GIDC.Plot No. 5227/A & BPhase 4VatvaAhmedabad - 382445,"65436946, 65456412",25840556 / 40085498,praghera@yahoo.co.in,www.darshiniengineers.com,Jignesh P. Aghera,9825901490,"Bell Conveyors, Coal Handling Plants, All Type of Industrial Equipments",N/A
+1408,DAYANA POLY PLAST,D-090/IV/A,Plastic Bags,2006,GIDC.Plot No. 2601/A-1Phase 4VatvaAhmedabad - 382445,25840387,25840387,dayanapolyplastpvtltd@gmail.com,N/A,Bharat A. Patel | Pragnesh A. Patel,9825493333 | 972779369,"HDPE Woven Bags, HDPE Laminated Fabrics, HDPE Woven Fabrics","HDPE Granules, LDPE Granules, LDPE Granules"
+1409,DEEP DYES & INTERMEDIATES PVT. LTD.,D-005/IV/A,"Chemical, Dyes, Pigments and Allied Industries",2003,GIDC.Plot No. C1-4520Phase 4VatvaAhmedabad - 382445,"25840973, 40085983",N/A,chiragdeep@hotmail.com,www.deepdyes.com,Chirag Vora | Swapan Vora,9825030431 | 9825777919,Reactive Dyes,"NMJ, STA"
+1410,DEEPAK DYES & CHEMICALS INDUSTRIES,D-006/I/L,Chemical Mfg and Suppliers,1978,"F Road , GIDC.Plot No. 77/5Phase 1VatvaAhmedabad - 382445",25830312,N/A,desaikiritp@gmail.com,N/A,K. P. Desai,9825559594,Auxillary (Textile Binder Main Product ),N/A
